feat: 쿠폰 모달 상태 분리 및 쿠폰 사용하기 버튼 기능 적용
